backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
deepgram-sdk==3.4.0
python-dotenv==1.0.0

# Video analysis (optional - only if using speaker sync)
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3